[codeql] Widen 16 int-multiplication-before-cast sites to avoid overflow
CodeQL's cpp/integer-multiplication-cast-to-long flagged 16 new sites
where two narrower operands (unsigned int, or a uint32_t returned by
align()/ceilDiv()) are multiplied and the unsigned-int result is then
implicitly converted to a wider type -- size_t, int64_t, or a vector's
size_type -- for an allocation size, buffer offset, or element count.
On dimensions large enough to overflow 32 bits, the multiply wraps
before the widening happens, so the destination sees a truncated value
instead of the true count.

Each site now casts one operand to the destination width before the
multiply: nntrainer/layers/llm/tie_word_embedding.cpp (per-row weight
offset), nntrainer/tensor/float_tensor.cpp (broadcast stride offset),
nntrainer/tensor/int4_utils.cpp (padded scales buffer size), and the
row/column products test/unittest/unittest_nntrainer_cpu_backend.cpp
builds vectors and calls quantization helpers with.

+++ test/unittest/unittest_nntrainer_cpu_backend.cpp
@@ -531,7 +531,7 @@ float test_gemm_q4_0(const uint32_t M, const uint32_t K, const uint32_t N,
 
   // Step3. Run GEMM! (Online activation quantization + kernel routine + return
   // float)
-  std::vector<float> dst(M * N);
+  std::vector<float> dst(static_cast<size_t>(M) * N);
   auto t1 = high_resolution_clock::now();
   // #### MAIN TESTED METHOD ####
   nntrainer::gemm_q4_0(M, N, K, activations, K,
@@ -574,7 +574,7 @@ float test_gemm_q4_K(const uint32_t M, const uint32_t K, const uint32_t N,
 
   // Step3. Run GEMM! (Online activation quantization + kernel routine + return
   // float)
-  std::vector<float> dst(M * N);
+  std::vector<float> dst(static_cast<size_t>(M) * N);
   auto t1 = high_resolution_clock::now();
   // #### MAIN TESTED METHOD ####
   nntrainer::gemm_q4_K(M, N, K, activations, K, (void *)repacked_qWeight.data(),
@@ -611,7 +611,7 @@ float test_gemm_q6_K(const uint32_t M, const uint32_t K, const uint32_t N,
 
   // Step2. Run GEMM! (Online activation quantization + kernel routine + return
   // float)
-  std::vector<float> dst(M * N);
+  std::vector<float> dst(static_cast<size_t>(M) * N);
   auto t1 = high_resolution_clock::now();
   // #### MAIN TESTED METHOD ####
   nntrainer::gemm_q6_K(M, N, K, activations, K, (void *)offline_qWeight_ptr, N,
@@ -643,9 +643,11 @@ static void run_quant_test(const uint32_t M, const uint32_t K, const uint32_t N,
   ///@note A(sizez, sizex) * W.T(sizey, sizex) = (sizez, sizey)
 
   ///@note q4_K GEMM is a Row-Major, transB GEMM
-  std::vector<float> activation = generate_random_vector<float>(M * K);
-  std::vector<float> weight = generate_random_vector<float>(N * K);
-  std::vector<float> ref_dst(M * N);
+  std::vector<float> activation =
+    generate_random_vector<float>(static_cast<size_t>(M) * K);
+  std::vector<float> weight =
+    generate_random_vector<float>(static_cast<size_t>(N) * K);
+  std::vector<float> ref_dst(static_cast<size_t>(M) * N);
 
   // GROUND TRUTH TRANSB SGEMM for reference
   auto t1 = high_resolution_clock::now();
@@ -1476,9 +1478,10 @@ static void run_transform_int4_test_(const uint32_t K, const uint32_t N,
   std::vector<float> weight_fp32;
   if (use_ones) {
     float ones_ratio = 0.1f;
-    weight_fp32 = generate_01_vector(N * K, ones_ratio);
+    weight_fp32 = generate_01_vector(static_cast<size_t>(N) * K, ones_ratio);
   } else {
-    weight_fp32 = generate_random_vector<float, false>(N * K, -1.0, 1.0);
+    weight_fp32 = generate_random_vector<float, false>(
+      static_cast<size_t>(N) * K, -1.0, 1.0);
   }
 
   bool print = false;
@@ -1510,9 +1513,10 @@ static void run_transform_int4_test_(const uint32_t K, const uint32_t N,
   std::vector<uint8_t> unpacked_weights_q4(q4_data_size);
   nntrainer::unpack_q4_0(dst_q4_0x.data(), unpacked_weights_q4.data(),
                          q4_data_size, N, K);
-  std::vector<float> dequantized_weights_q4(N * K);
+  std::vector<float> dequantized_weights_q4(static_cast<size_t>(N) * K);
   nntrainer::dequantize_row_q4_0(unpacked_weights_q4.data(),
-                                 dequantized_weights_q4.data(), N * K);
+                                 dequantized_weights_q4.data(),
+                                 static_cast<int64_t>(N) * K);
   if (print && use_ones) {
     printMatrixI("dequantized_weights_q4 I", dequantized_weights_q4.data(), N,
                  K);
@@ -1525,7 +1529,7 @@ static void run_transform_int4_test_(const uint32_t K, const uint32_t N,
 
   // Reference solution - Int4 data (osv32_isv2) --> FP32 --> quantization to
   // Q4_0x8 For checking difference of accuracy
-  std::vector<float> dequant_weight_fp32(N * K);
+  std::vector<float> dequant_weight_fp32(static_cast<size_t>(N) * K);
   nntrainer::Int4Utils::dequantizePacked(osv32_weights, osv32_scales, N, K,
                                          scale_group_size, dequant_weight_fp32);
   std::vector<uint8_t> tmp_q4_weight(q4_data_size);
@@ -1538,9 +1542,10 @@ static void run_transform_int4_test_(const uint32_t K, const uint32_t N,
   std::vector<uint8_t> unpacked_ref_weights_q4(q4_data_size);
   nntrainer::unpack_q4_0(ref_q4_0x8.data(), unpacked_ref_weights_q4.data(),
                          q4_data_size, N, K);
-  std::vector<float> dequantized_ref_weights_q4(N * K);
+  std::vector<float> dequantized_ref_weights_q4(static_cast<size_t>(N) * K);
   nntrainer::dequantize_row_q4_0(unpacked_ref_weights_q4.data(),
-                                 dequantized_ref_weights_q4.data(), N * K);
+                                 dequantized_ref_weights_q4.data(),
+                                 static_cast<int64_t>(N) * K);
   float mse_fp32_transform_q4 =
     mse<float>(weight_fp32.data(), dequantized_ref_weights_q4.data(), N * K);
   std::cout << "MSE FP32 transform Q4_0:   " << std::setprecision(10)
